Makefileを更新し、MLモデルのトレーニングスクリプトのパスを`src/ml/training/`に変更しました。また、不要なモデル関連ファイルを削除し、`ml_report.py`を`src/ml/training/`に移動しました。これにより、コードの整理と一貫性が向上しました。
