+++ src/artifacts/ml_results.csv
@@ -1,0 +1,12 @@
+timestamp,model,rmse,mae,r2
+2025-06-28 17:55:37,xgb,1257.7152025667542,344.4248511904762,0.9999875900233295
+2025-06-28 17:55:43,xgb,1257.7152025667542,344.4248511904762,0.9999875900233295
+2025-06-28 17:55:45,cat,11360.07200404268,8389.523271778236,0.9989875618249385
+2025-06-28 17:55:45,lgbm,59694.15940371752,32171.354846639868,0.9720443222219067
+2025-06-28 17:56:24,ensemble,22286.456392730895,12003.65105581824,0.9961033770499316
+2025-06-28 17:56:34,stacking,22286.456392730895,12003.65105581824,0.9961033770499316
+2025-06-28 17:56:53,xgb,1257.7152025667542,344.4248511904762,0.9999875900233295
+2025-06-28 17:56:54,cat,11360.07200404268,8389.523271778236,0.9989875618249385
+2025-06-28 17:56:55,lgbm,59694.15940371752,32171.354846639868,0.9720443222219067
+2025-06-28 17:56:56,ensemble,22286.456392730895,12003.65105581824,0.9961033770499316
+2025-06-28 17:56:57,stacking,22286.456392730895,12003.65105581824,0.9961033770499316
