+++ src/ml/__init__.py
@@ -1,0 +1,14 @@
+"""
+ML (Machine Learning) module for solid-ml-stack
+
+This module contains all machine learning related functionality:
+- models: Model definitions and implementations
+- training: Training and inference pipelines
+- utils: ML-specific utilities
+"""
+
+from . import models
+from . import training
+from . import utils
+
+__all__ = ['models', 'training', 'utils']
