+++ src/ml/models/lgbm_model.py
@@ -1,31 +1,31 @@
 import sys
 import os
 
-# srcディレクトリとmodelsディレクトリをPYTHONPATHに追加
+# srcディレクトリをPYTHONPATHに追加
+sys.path.insert(0, os.path.join(os.path.dirname(__file__), '..', '..'))
 sys.path.insert(0, os.path.join(os.path.dirname(__file__), '..'))
-sys.path.insert(0, os.path.dirname(__file__))
 
-from catboost import CatBoostRegressor
+import lightgbm as lgb
 import numpy as np
 import pandas as pd
 from typing import Any, Union, Optional
-from base_model import BaseModel, ModelConfig
-from utils.config import CatBoostConfig
+from .base_model import BaseModel, ModelConfig
+from utils.config import LightGBMConfig
 from utils.base import set_seed
 
 
-class CatBoostModel(BaseModel):
-    """CatBoostモデルクラス"""
+class LightGBMModel(BaseModel):
+    """LightGBMモデルクラス"""
     
-    def __init__(self, config: Optional[ModelConfig] = None, cat_config: Optional[CatBoostConfig] = None):
+    def __init__(self, config: Optional[ModelConfig] = None, lgbm_config: Optional[LightGBMConfig] = None):
         super().__init__(config)
-        self.cat_config = cat_config or CatBoostConfig()
+        self.lgbm_config = lgbm_config or LightGBMConfig()
     
     def get_model_name(self) -> str:
-        return "CatBoost"
+        return "LightGBM"
     
     def _train(self, X: Union[np.ndarray, pd.DataFrame], y: Union[np.ndarray, pd.Series]) -> Any:
-        """CatBoostモデルを学習"""
+        """LightGBMモデルを学習"""
         set_seed(self.config.seed)
         
         # --- データを確実に数値に変換 ---
@@ -46,24 +46,30 @@ class CatBoostModel(BaseModel):
             # 全ての列をfloat型に変換
             X = X_df.astype(float).values
         
+        # Dataset作成
+        dtrain = lgb.Dataset(X, label=y)
+        
         # パラメータ設定
         params = {
-            "iterations": self.cat_config.iterations,
-            "depth": self.cat_config.depth,
-            "learning_rate": self.cat_config.learning_rate,
-            "loss_function": self.cat_config.loss_function,
-            "verbose": self.cat_config.verbose,
-            "random_seed": self.config.seed
+            "objective": self.lgbm_config.objective,
+            "seed": self.config.seed,
+            "max_depth": self.lgbm_config.max_depth,
+            "learning_rate": self.lgbm_config.learning_rate,
+            "subsample": self.lgbm_config.subsample,
+            "colsample_bytree": self.lgbm_config.colsample_bytree,
+            "verbosity": self.lgbm_config.verbosity,
+            "num_leaves": 31,
+            "min_child_samples": 20,
+            "reg_alpha": 0.0,
+            "reg_lambda": 0.0
         }
         
-        # モデル作成と学習
-        model = CatBoostRegressor(**params)
-        model.fit(X, y)
-        
+        # モデル学習
+        model = lgb.train(params, dtrain, num_boost_round=self.lgbm_config.num_boost_round)
         return model
     
     def _predict(self, model: Any, X: Union[np.ndarray, pd.DataFrame]) -> np.ndarray:
-        """CatBoostモデルで予測"""
+        """LightGBMモデルで予測"""
         # --- データを確実に数値に変換 ---
         if isinstance(X, pd.DataFrame):
             X = X.copy()  # 元データを変更しないようコピー
@@ -84,13 +90,13 @@ class CatBoostModel(BaseModel):
         
         return model.predict(X)
     
-    def get_feature_importance(self) -> dict:
+    def get_feature_importance(self, importance_type: str = "gain") -> dict:
         """特徴量重要度を取得"""
         if not self.is_fitted or self.model is None:
             raise ValueError("Model is not fitted")
         
-        importance = self.model.get_feature_importance()
-        feature_names = self.model.feature_names_
+        importance = self.model.feature_importance(importance_type=importance_type)
+        feature_names = self.model.feature_name()
         
         if feature_names is None:
             feature_names = [f"feature_{i}" for i in range(len(importance))]
@@ -98,16 +104,10 @@ class CatBoostModel(BaseModel):
         importance_dict = dict(zip(feature_names, importance))
         return dict(sorted(importance_dict.items(), key=lambda x: x[1], reverse=True))
     
-    def get_feature_importance_type(self, importance_type: str = "LossFunctionChange") -> dict:
-        """指定されたタイプの特徴量重要度を取得"""
-        if not self.is_fitted or self.model is None:
-            raise ValueError("Model is not fitted")
-        
-        importance = self.model.get_feature_importance(type=importance_type)
-        feature_names = self.model.feature_names_
-        
-        if feature_names is None:
-            feature_names = [f"feature_{i}" for i in range(len(importance))]
-        
-        importance_dict = dict(zip(feature_names, importance))
-        return dict(sorted(importance_dict.items(), key=lambda x: x[1], reverse=True)) 
+    def get_feature_importance_split(self) -> dict:
+        """分割ベースの特徴量重要度を取得"""
+        return self.get_feature_importance(importance_type="split")
+    
+    def get_feature_importance_gain(self) -> dict:
+        """ゲインベースの特徴量重要度を取得"""
+        return self.get_feature_importance(importance_type="gain")
