+++ src/ml/training/__init__.py
@@ -1,0 +1,10 @@
+"""
+ML Training module
+
+Contains training and inference pipelines for machine learning models.
+"""
+
+from .train import main as train_main
+from .ml_report import main as report_main
+
+__all__ = ['train_main', 'report_main']
