+++ src/ml/utils/__init__.py
@@ -1,0 +1,9 @@
+"""
+ML Utilities module
+
+Contains utility functions for machine learning operations.
+"""
+
+from .ml_report import get_latest_results, main as report_main
+
+__all__ = ['get_latest_results', 'report_main']
